Switched off Actor Modeling
Rather than Actors without guarantees switching to a SlateRequest approach with fully functional approach.
