--- src/main/scala/edu/eckerd/integrations/slate/housing/application/models/HousingRequestResponse.scala
@@ -1,8 +1,0 @@
-package edu.eckerd.integrations.slate.housing.application.models
-
-/**
-  * Created by davenpcm on 6/17/16.
-  */
-case class HousingRequestResponse(
-                                 row: List[HousingRequest]
-                                 )
